Broken sampling code but has a nice button
