+++ src/database/repository/preset.ts
@@ -14,6 +14,7 @@ export interface DatabasePreset {
 }
 
 export const PresetRepository = dataSource.getRepository(Preset).extend({
+	//get one preset
 	getItem(key: string): string | null {
 		const item = this.findOne({
 			select: {
@@ -26,6 +27,7 @@ export const PresetRepository = dataSource.getRepository(Preset).extend({
 		if (!item) return null
 		else return item.value
 	},
+	//get all presets
 	async getAll(): Promise<Array<DatabasePreset>> {
 		const items = await this.find()
 		return items.map((item: Preset) => {
@@ -41,6 +43,7 @@ export const PresetRepository = dataSource.getRepository(Preset).extend({
 			}
 		})
 	},
+	//replace all presets
 	async setAll(presets: Array<DatabasePreset>): Promise<void> {
 		await this.delete({})
 		// Convert preset back to an object
@@ -54,5 +57,9 @@ export const PresetRepository = dataSource.getRepository(Preset).extend({
 			}
 		})
 		await this.insert(presetsToInsert)
+	},
+	//insert one preset - wrapped with correct type
+	async insertOne(preset: DatabasePreset): Promise<void> {
+		await this.insert(preset)
 	},
 })
