+++ src/output/e131/constructor.ts
@@ -9,7 +9,9 @@ export const createE131 = async () => {
 		const sourceName = await ConfigRepository.getItem('e131SourceName')
 		const priority = parseInt(await ConfigRepository.getItem('e131Priority'))
 		const frequency = parseInt(await ConfigRepository.getItem('e131Frequency'))
-		globalThis.e131 = new E131(firstUniverse, universes, sourceName, priority, frequency)
+		const effectMode = (await ConfigRepository.getItem('e131Sampler_effectMode')) == '0' ? false : true
+		const sampleTime = parseInt(await ConfigRepository.getItem('e131Sampler_time'))
+		globalThis.e131 = new E131(firstUniverse, universes, sourceName, priority, frequency, effectMode, sampleTime)
 	}
 }
 export const destroyE131 = (): Promise<void> => {
