+++ src/output/e131/index.ts
@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/no-this-alias */
 import { networkInterfaces } from 'os'
 import e131Lib from '@paradise-pi/e131'
+import { PresetRepository } from '../../database/repository/preset'
+import { broadcast } from '../../api/broadcast'
 
 interface channelData {
 	channel: number
@@ -15,6 +17,14 @@ interface channelFade {
 	fadeFromTimestamp: number
 	fadeToTimestamp: number
 }
+
+interface UniverseData {
+	[sourceName: string]: {
+		[universe: number]: {
+			[channel: number]: number
+		}
+	}
+}
 /**
  * sACN = E1.31
  */
@@ -28,13 +38,25 @@ export class E131 {
 	protected priority: number
 	protected frequency: number
 	protected running: boolean
+	private effectMode: boolean
+	private sampleTime: number
 
-	constructor(firstUniverse: number, universes: number, sourceName: string, priority: number, frequency: number) {
+	constructor(
+		firstUniverse: number,
+		universes: number,
+		sourceName: string,
+		priority: number,
+		frequency: number,
+		effectMode: boolean,
+		sampleTime: number
+	) {
 		this.firstUniverse = firstUniverse
 		this.universes = universes
 		this.sourceName = sourceName
 		this.priority = priority
 		this.frequency = frequency
+		this.effectMode = effectMode
+		this.sampleTime = sampleTime
 		this.running = true
 		this.setupUniverses()
 		this.initSending()
@@ -202,5 +224,112 @@ export class E131 {
 	private getUniqueId() {
 		const interfaces = networkInterfaces()
 		return Buffer.from(interfaces[Object.keys(interfaces)[0]][0].mac) // Get a unique ID to set as the CID
+	}
+
+	public async sampleE131() {
+		const delay = (time: number) => {
+			return new Promise(res => {
+				setTimeout(res, time)
+			})
+		}
+		await this.terminate()
+		logger.info('Starting Sampling Mode')
+		//log effect mode
+		logger.info(
+			this.effectMode
+				? 'Storing first value of a varying value (EFFECT MODE ON)'
+				: 'Ignoring varying values (EFFECT MODE OFF)'
+		)
+
+		//how many universes are we sampling?
+		//20 universe limit is enforced due to memory limitations
+		const numUniverses = this.universes > 20 ? 20 : this.universes
+
+		const universes = []
+		for (let i = 1; i <= numUniverses; i++) {
+			universes.push(i)
+		}
+		const universeData: UniverseData = {}
+
+		const server = new e131Lib.Server(universes)
+		logger.debug('E1.31 Server started')
+
+		server.on('listening', () => {
+			logger.info('Listening on ') //port ' + server.port + '- universes ' + server.universes)
+		})
+
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		server.on('packet', (packet: any) => {
+			const sourceName: string = packet.getSourceName().replace(/\0/g, '')
+			const universe: number = packet.getUniverse()
+			const slotsData = packet.getSlotsData()
+			const priority: number = packet.getPriority()
+
+			if (priority != 0) {
+				// For some reason, known only to the developers of this library/sACN (I'm not even sure)........you get some bizarre packets that are just priorities and nothing else from time to time. The only feature of these I can find is that they have no priority, so if you find one without priority just ignore it and hope for the best.
+				if (universeData[sourceName] === undefined) {
+					universeData[sourceName] = {}
+					logger.info('Found new device ' + sourceName)
+				}
+				if (universeData[sourceName][universe] === undefined) {
+					universeData[sourceName][universe] = {}
+					logger.info('Found universe ' + universe + ' for device ' + sourceName)
+				}
+				for (let i = 0; i < slotsData.length; i++) {
+					if (universeData[sourceName][universe][i + 1] === undefined) {
+						universeData[sourceName][universe][i + 1] = slotsData[i]
+					} else if (!this.effectMode && universeData[sourceName][universe][i + 1] === null) {
+						// This has already been marked as jittery - so ignore it
+					} else if (!this.effectMode && universeData[sourceName][universe][i + 1] !== slotsData[i]) {
+						// So we've found data that doesn't match what we had down for it before, so it might be that an effect is running. The best way to deal with this is to mark it as false, which means it won't be saved (on purpose)
+						universeData[sourceName][universe][i + 1] = null
+						logger.warn(
+							'Discarding data for channel ' +
+								i +
+								' due to value change (universe ' +
+								universe +
+								' from device ' +
+								sourceName +
+								') - is an effect running?'
+						)
+					}
+				}
+			}
+		})
+
+		//Timeout timer
+		let timeoutTimerDuration = 15000 // 15 seconds is the default
+		if (!(this.sampleTime === undefined || this.sampleTime > 300 || this.sampleTime < 5)) {
+			timeoutTimerDuration = this.sampleTime * 1000
+		}
+
+		await delay(timeoutTimerDuration).then(() => {
+			server.close()
+			logger.debug('Finished Sampling')
+			for (const [deviceName, device] of Object.entries(universeData)) {
+				for (const [universeID, universeData] of Object.entries(device)) {
+					for (const key in universeData) {
+						// eslint-disable-next-line no-prototype-builtins
+						if (universeData.hasOwnProperty(key) && universeData[key] === null) {
+							delete universeData[key] // Remove null values
+						}
+					}
+					PresetRepository.insertOne({
+						name: 'Universe ' + universeID + ' sampled from ' + deviceName,
+						enabled: true,
+						universe: parseInt(universeID),
+						data: JSON.parse(JSON.stringify(universeData)),
+					})
+				}
+			}
+			logger.info('Resuming E1.31 Connection')
+			this.initSending()
+		})
+
+		const timeoutStarted = +new Date()
+		setInterval(function () {
+			const currentTime = +new Date()
+			broadcast('progress', { progress: currentTime - timeoutStarted, total: timeoutTimerDuration })
+		}, 500)
 	}
 }
